requirements.txt
flask
requests
flask-wtf
python-dotenv
